+++ .eslintrc.json
@@ -40,7 +40,6 @@
         "no-console": ["error", { "allow": ["error"] }],
         "no-undef": "off",
         "semi": ["error", "always"],
-        "@typescript-eslint/no-unused-vars": "error",
         "sonarjs/cognitive-complexity": "warn"
       }
     },
